Update package versions in project files
Updated the version of the `Azure.Data.Tables` package from `12.10.0` to `12.11.0` in the `L4D2PlayStats.Core.csproj` file.
Updated the version of the `FluentValidation` package from `11.11.0` to `12.0.0` in the `L4D2PlayStats.Core.csproj` file.
Updated the version of the `FluentValidation.DependencyInjectionExtensions` package from `11.11.0` to `12.0.0` in the `L4D2PlayStats.FunctionApp.csproj` file.